Updated the front end
Updated front end so the background and font color would change depending on who wins. Also added a function that will automatically copy the new index.html, and assets folder if the front end were to be changed you need to do this or it wont update. To automatically copy the files over go into the terminal for BetAI and then go to cd react for the react folder terminal and run the code npm run build